Bring constants into own file
