+++ src/js/constants/ServiceActionItem.js
@@ -1,0 +1,9 @@
+const ServiceActionItem = {
+  EDIT: 'edit',
+  DESTROY: 'destroy',
+  SCALE: 'scale',
+  SUSPEND: 'suspend',
+  MORE: 'more'
+};
+
+module.exports = ServiceActionItem;
